Mobile GRAI picking: remove invalid scan-time PO-reference match
The scan-time check refused a returnable-asset GRAI whose serial did not
match the current sales order's PO reference. It is invalid: those GRAIs are
dummy GRAIs the system generates at shipment with the serial derived from the
order's PO reference, so at scan time there is nothing authoritative to
validate against. Remove the check, its PO-reference plumbing, the
GRAIPOReferenceMismatch message, and the DummyGRAITemplate match helpers.

The LU-wide GRAI dedupe is unaffected. Replace the two PO-reference-match
mobile E2E specs with one asserting a GRAI not matching the order's PO
reference is now accepted and the TU resolves.
