+++ backend/de.metas.handlingunits.base/src/main/java/de/metas/handlingunits/grai/DummyGRAITemplate.java
@@ -64,43 +64,6 @@ public class DummyGRAITemplate
 		}
 	}
 
-	/**
-	 * @return {@code true} iff {@code grai} is non-null and its canonical value starts with this template's
-	 * {@code companyPrefix.assetType.serialPrefix} prefix — i.e. it was (or could have been) generated by this
-	 * template's {@link #buildGRAI(int)}.
-	 */
-	public boolean matches(@Nullable final GRAI grai)
-	{
-		if (grai == null)
-		{
-			return false;
-		}
-		return grai.toCanonicalString().startsWith(prefix());
-	}
-
-	/**
-	 * @return {@code true} iff {@code grai} has the Migros returnable-asset structure, i.e. its companyPrefix and
-	 * assetType equal {@link #MIGROS_COMPANY_PREFIX}/{@link #MIGROS_ASSET_TYPE}. Does NOT check the serial/PO
-	 * reference — use {@link #matches(GRAI)} for that. Never throws; a non-Migros or malformed GRAI yields
-	 * {@code false}.
-	 */
-	public static boolean isMigrosStructure(@Nullable final GRAI grai)
-	{
-		if (grai == null)
-		{
-			return false;
-		}
-		try
-		{
-			return MIGROS_COMPANY_PREFIX.equals(grai.getCompanyPrefix())
-					&& MIGROS_ASSET_TYPE.equals(grai.getAssetType());
-		}
-		catch (final RuntimeException e)
-		{
-			return false;
-		}
-	}
-
 	@NonNull
 	private String prefix()
 	{
